plastic-identifier/scripts/requirements.txt
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
comm==0.1.2
contourpy==1.0.6
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
fonttools==4.38.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
protobuf==4.21.11
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.6
pyzmq==25.0.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sv-ttk==2.4
threadpoolctl==3.1.0
tk==0.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
wcwidth==0.2.6
Werkzeug==2.2.3
